Adicionado Testes

+++ src/main/java/br/com/edgarneto/medicalconsult/consulta/models/Consulta.java
@@ -20,7 +20,7 @@ public class Consulta {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_CONSULTA")
-    private Integer idConsulta;
+    private Long idConsulta;
     @Column(name = "DATA_CONSULTA")
     private Date dataConsulta;
     @Column(name = "PROFISSIONAL")
